fix: markdown-mode for *scratch* errors
Setting markdown mode for scratch was throwing errors because the markdown-mode
didn't exist yet.  Moving that config to the bottom of the config file caused a
small visual glitch on startup.  Rather than figure out how to order the loading
to make the visual glitch go away, I am moving the markdown mode to a higher
location in the config and setting the scratch mode in the use-package
invocation for markdown-mode.  Updated comments accordingly.